+++ .snapshots/successful/adt_triple/compiler/compiled.js
@@ -53,8 +53,8 @@ function v_h0(v_a, v_b, v_c){
       const __s = __concat(v_a, " ");
       switch (__s[0]) {
         case 3: {
-          const v__do_e_20_3 = __s[1];
-          return [3, v__do_e_20_3];
+          const v__do_e_2 = __s[1];
+          return [3, v__do_e_2];
         }
         case 4: {
           const v_s0 = __s[1];
@@ -62,8 +62,8 @@ function v_h0(v_a, v_b, v_c){
             const __s = __concat(v_s0, v_b);
             switch (__s[0]) {
               case 3: {
-                const v__do_e_21_3 = __s[1];
-                return [3, v__do_e_21_3];
+                const v__do_e_1 = __s[1];
+                return [3, v__do_e_1];
               }
               case 4: {
                 const v_s1 = __s[1];
@@ -71,8 +71,8 @@ function v_h0(v_a, v_b, v_c){
                   const __s = __concat(v_s1, " ");
                   switch (__s[0]) {
                     case 3: {
-                      const v__do_e_22_3 = __s[1];
-                      return [3, v__do_e_22_3];
+                      const v__do_e_0 = __s[1];
+                      return [3, v__do_e_0];
                     }
                     case 4: {
                       const v_s2 = __s[1];