+++ .snapshots/successful/polymorphism_function-different-types/compiler/compiled.js
@@ -56,7 +56,7 @@ function v_showResult(v_r){
     }
 }
 
-const main = (v__let_12)(((s) => { switch(s[0]) { case 3: { const v__do_e_21_9 = s[1]; return [3, v__do_e_21_9]; } case 4: { const v_s0 = s[1]; return __concat(v_s0, (v_showResult)((v_unwrap)([20, [21, "nested"]]))); } } })(__concat((v_unwrap)([20, "from box"]), " ")));
+const main = (v__let_12)(((s) => { switch(s[0]) { case 3: { const v__do_e_0 = s[1]; return [3, v__do_e_0]; } case 4: { const v_s0 = s[1]; return __concat(v_s0, (v_showResult)((v_unwrap)([20, [21, "nested"]]))); } } })(__concat((v_unwrap)([20, "from box"]), " ")));
 
 function v__let_12(v_res){
     {