+++ .snapshots/successful/strings_split-onfirst/compiler/compiled.js
@@ -62,8 +62,8 @@ function v_renderTuple(v_a, v_b){
       const __s = __concat("Just(", v_a);
       switch (__s[0]) {
         case 3: {
-          const v__do_e_13_3 = __s[1];
-          return [3, v__do_e_13_3];
+          const v__do_e_2 = __s[1];
+          return [3, v__do_e_2];
         }
         case 4: {
           const v_s0 = __s[1];
@@ -71,8 +71,8 @@ function v_renderTuple(v_a, v_b){
             const __s = __concat(v_s0, "|");
             switch (__s[0]) {
               case 3: {
-                const v__do_e_14_3 = __s[1];
-                return [3, v__do_e_14_3];
+                const v__do_e_1 = __s[1];
+                return [3, v__do_e_1];
               }
               case 4: {
                 const v_s1 = __s[1];
@@ -80,8 +80,8 @@ function v_renderTuple(v_a, v_b){
                   const __s = __concat(v_s1, v_b);
                   switch (__s[0]) {
                     case 3: {
-                      const v__do_e_15_3 = __s[1];
-                      return [3, v__do_e_15_3];
+                      const v__do_e_0 = __s[1];
+                      return [3, v__do_e_0];
                     }
                     case 4: {
                       const v_s2 = __s[1];
@@ -97,7 +97,7 @@ function v_renderTuple(v_a, v_b){
     }
 }
 
-const main = (v__let_12)(((s) => { switch(s[0]) { case 3: { const v__do_e_21_9 = s[1]; return [3, v__do_e_21_9]; } case 4: { const v_a = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_22_9 = s[1]; return [3, v__do_e_22_9]; } case 4: { const v_b = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_23_9 = s[1]; return [3, v__do_e_23_9]; } case 4: { const v_c = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_24_9 = s[1]; return [3, v__do_e_24_9]; } case 4: { const v_d = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_25_9 = s[1]; return [3, v__do_e_25_9]; } case 4: { const v_e = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_26_9 = s[1]; return [3, v__do_e_26_9]; } case 4: { const v_f = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_27_9 = s[1]; return [3, v__do_e_27_9]; } case 4: { const v_g = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_28_9 = s[1]; return [3, v__do_e_28_9]; } case 4: { const v_h = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_29_9 = s[1]; return [3, v__do_e_29_9]; } case 4: { const v_r0 = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_30_9 = s[1]; return [3, v__do_e_30_9]; } case 4: { const v_r1 = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_31_9 = s[1]; return [3, v__do_e_31_9]; } case 4: { const v_r2 = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_32_9 = s[1]; return [3, v__do_e_32_9]; } case 4: { const v_r3 = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_33_9 = s[1]; return [3, v__do_e_33_9]; } case 4: { const v_r4 = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_34_9 = s[1]; return [3, v__do_e_34_9]; } case 4: { const v_r5 = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_35_9 = s[1]; return [3, v__do_e_35_9]; } case 4: { const v_r6 = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_36_9 = s[1]; return [3, v__do_e_36_9]; } case 4: { const v_r7 = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_37_9 = s[1]; return [3, v__do_e_37_9]; } case 4: { const v_r8 = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_38_9 = s[1]; return [3, v__do_e_38_9]; } case 4: { const v_r9 = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_39_9 = s[1]; return [3, v__do_e_39_9]; } case 4: { const v_r10 = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_40_9 = s[1]; return [3, v__do_e_40_9]; } case 4: { const v_r11 = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_41_9 = s[1]; return [3, v__do_e_41_9]; } case 4: { const v_r12 = s[1]; return __concat(v_r12, v_h); } } })(__concat(v_r11, ", ")); } } })(__concat(v_r10, v_g)); } } })(__concat(v_r9, ", ")); } } })(__concat(v_r8, v_f)); } } })(__concat(v_r7, ", ")); } } })(__concat(v_r6, v_e)); } } })(__concat(v_r5, ", ")); } } })(__concat(v_r4, v_d)); } } })(__concat(v_r3, ", ")); } } })(__concat(v_r2, v_c)); } } })(__concat(v_r1, ", ")); } } })(__concat(v_r0, v_b)); } } })(__concat(v_a, ", ")); } } })((v_render)(__splitOnFirst("abcde", "ab"))); } } })((v_render)(__splitOnFirst("abc", "abc"))); } } })((v_render)(__splitOnFirst(":", "foo:"))); } } })((v_render)(__splitOnFirst(":", ":foo"))); } } })((v_render)(__splitOnFirst("", "abc"))); } } })((v_render)(__splitOnFirst("x", "abc"))); } } })((v_render)(__splitOnFirst("::", "user::42::admin"))); } } })((v_render)(__splitOnFirst(",", "a,b,c"))));
+const main = (v__let_12)(((s) => { switch(s[0]) { case 3: { const v__do_e_23 = s[1]; return [3, v__do_e_23]; } case 4: { const v_a = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_22 = s[1]; return [3, v__do_e_22]; } case 4: { const v_b = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_21 = s[1]; return [3, v__do_e_21]; } case 4: { const v_c = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_20 = s[1]; return [3, v__do_e_20]; } case 4: { const v_d = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_19 = s[1]; return [3, v__do_e_19]; } case 4: { const v_e = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_18 = s[1]; return [3, v__do_e_18]; } case 4: { const v_f = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_17 = s[1]; return [3, v__do_e_17]; } case 4: { const v_g = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_16 = s[1]; return [3, v__do_e_16]; } case 4: { const v_h = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_15 = s[1]; return [3, v__do_e_15]; } case 4: { const v_r0 = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_14 = s[1]; return [3, v__do_e_14]; } case 4: { const v_r1 = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_13 = s[1]; return [3, v__do_e_13]; } case 4: { const v_r2 = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_12 = s[1]; return [3, v__do_e_12]; } case 4: { const v_r3 = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_11 = s[1]; return [3, v__do_e_11]; } case 4: { const v_r4 = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_10 = s[1]; return [3, v__do_e_10]; } case 4: { const v_r5 = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_9 = s[1]; return [3, v__do_e_9]; } case 4: { const v_r6 = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_8 = s[1]; return [3, v__do_e_8]; } case 4: { const v_r7 = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_7 = s[1]; return [3, v__do_e_7]; } case 4: { const v_r8 = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_6 = s[1]; return [3, v__do_e_6]; } case 4: { const v_r9 = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_5 = s[1]; return [3, v__do_e_5]; } case 4: { const v_r10 = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_4 = s[1]; return [3, v__do_e_4]; } case 4: { const v_r11 = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_3 = s[1]; return [3, v__do_e_3]; } case 4: { const v_r12 = s[1]; return __concat(v_r12, v_h); } } })(__concat(v_r11, ", ")); } } })(__concat(v_r10, v_g)); } } })(__concat(v_r9, ", ")); } } })(__concat(v_r8, v_f)); } } })(__concat(v_r7, ", ")); } } })(__concat(v_r6, v_e)); } } })(__concat(v_r5, ", ")); } } })(__concat(v_r4, v_d)); } } })(__concat(v_r3, ", ")); } } })(__concat(v_r2, v_c)); } } })(__concat(v_r1, ", ")); } } })(__concat(v_r0, v_b)); } } })(__concat(v_a, ", ")); } } })((v_render)(__splitOnFirst("abcde", "ab"))); } } })((v_render)(__splitOnFirst("abc", "abc"))); } } })((v_render)(__splitOnFirst(":", "foo:"))); } } })((v_render)(__splitOnFirst(":", ":foo"))); } } })((v_render)(__splitOnFirst("", "abc"))); } } })((v_render)(__splitOnFirst("x", "abc"))); } } })((v_render)(__splitOnFirst("::", "user::42::admin"))); } } })((v_render)(__splitOnFirst(",", "a,b,c"))));
 
 function v__let_12(v_res){
     {