+++ .snapshots/successful/syntax_case-on-function-call/compiler/compiled.js
@@ -37,8 +37,8 @@ function v_search(v_key){
       const __s = __concat("found:", v_key);
       switch (__s[0]) {
         case 3: {
-          const v__do_e_9_3 = __s[1];
-          return [3, v__do_e_9_3];
+          const v__do_e_0 = __s[1];
+          return [3, v__do_e_0];
         }
         case 4: {
           const v_found = __s[1];