+++ .snapshots/successful/syntax_top-level-lambda-rhs/compiler/compiled.js
@@ -44,7 +44,7 @@ function v_curryFst(v_a, v_b){
     return v_a;
 }
 
-const main = (v__let_12)(((s) => { switch(s[0]) { case 3: { const v__do_e_21_9 = s[1]; return [3, v__do_e_21_9]; } case 4: { const v_s0 = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_22_9 = s[1]; return [3, v__do_e_22_9]; } case 4: { const v_s1 = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_23_9 = s[1]; return [3, v__do_e_23_9]; } case 4: { const v_s2 = s[1]; return __concat(v_s2, String((v_curryFst)((3|0), (9|0)))); } } })(__concat(v_s1, " ")); } } })(__concat(v_s0, String((v_constFst)((7|0), (99|0))))); } } })(__concat(String((v_f)((42|0))), " ")));
+const main = (v__let_12)(((s) => { switch(s[0]) { case 3: { const v__do_e_2 = s[1]; return [3, v__do_e_2]; } case 4: { const v_s0 = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_1 = s[1]; return [3, v__do_e_1]; } case 4: { const v_s1 = s[1]; return ((s) => { switch(s[0]) { case 3: { const v__do_e_0 = s[1]; return [3, v__do_e_0]; } case 4: { const v_s2 = s[1]; return __concat(v_s2, String((v_curryFst)((3|0), (9|0)))); } } })(__concat(v_s1, " ")); } } })(__concat(v_s0, String((v_constFst)((7|0), (99|0))))); } } })(__concat(String((v_f)((42|0))), " ")));
 
 function v__let_12(v_res){
     {